+++ src/main/java/de/markusbordihn/easymobfarm/block/MobFarmBlock.java
@@ -22,6 +22,7 @@ package de.markusbordihn.easymobfarm.block;
 import java.util.Collections;
 import java.util.List;
 import java.util.Set;
+
 import javax.annotation.Nullable;
 
 import net.minecraft.ChatFormatting;